+++ test/hexapdf/type/acro_form/test_form.rb
@@ -152,6 +152,12 @@ describe HexaPDF::Type::AcroForm::Form do
       assert(field.password_field?)
     end
 
+    it "creates a file select field" do
+      field = @acro_form.create_file_select_field("field")
+      assert_equal(:Tx, field.field_type)
+      assert(field.file_select_field?)
+    end
+
     it "creates a check box" do
       field = @acro_form.create_check_box("field")
       assert(field.check_box?)